requirements.txt
pdf2image
python-pptx

